osd: fix crash when hard resetting in the debugger
This patch fixes a crash when hard resetting a game in the debugger.
This patch introduces a new hook stop_debugger(), aimed to suspend
the Qt event loop processing during a shot period of time, between
the destruction of the debugger_manager object (when the machine
is recreated in mame.cpp), and when init_debugger() is called again.

A supplementary crash is fixed in init_debugger() when iterating and
deleting all widgets from topLevelWidgets(), because the widget list
may contains children of an already deleted parent widget.