thomson: relax specification of to9 internal drive
This drive is single sided, however we consider it like a (dual sided)
floppy_35_dd image, since the side one disablement is done when
handling the control register of the controller. This allows us to
accept dual sided floppy images even if we can only read and write on
the first side.