wd_fdc: increase the size of the flux write buffer
The write_flux() function is expensive because it iterates on the whole
track size in "remove damaged" and "write wspans" functions, so an
optimization is to increase the size of the write buffer to invoke them
less frequently. 2500 bytes if a value chosen to contains a full sector
of flux updates, so there is approximately 10-20 calls to write_flux()
during the formatting of a single track (the most expensive operation
from this pov). We don't loose information at the end of the pll write
operation, because they are concluded by a specific commit from function
stop_writing().