wd177x_dsk: optimize the format compatibility checks when saving
Checking the format compatibility for a given bitstream is an expensive
operation, because all tracks are processed against each candidate
format, and every sectors content is extracted by processing FM/MFM
format.

These are the same operations that'll be redone with the chosen format
in extract_sectors() later.

Even if we theorically allow a different format structure per track, an
optimization is to check just a couple of tracks (for example the first
track, the last track, and another in the middle of the disk).