wd177x_dsk: give up save when no candidate format exists
Using the wrong track format when no candidate has been identified is
often a cause a segfault later in this function, so it is better to give
up and not save anything. In this case, we end up with an image file
with size zero, but at least mame did not crash.