floppy: make weak bits less frequent
The problem with a 50% probability for a transition change is that it
becomes relatively easy to generate a FM sync code $f57e in these area,
that will confuse the controller later on this track.  Lowering the
probability to 25% only, still generates random weak bits, but with
statistically  less "1", and the sync code occurence is much less
frequent.

(MFM doesn't have this problem because three consecutive sync codes are
expected)

This scenario occurs with the Thomson CD90-015 controller based on a
MC6843 FDC. The BIOS code formatting a track first iterates over all
sector headers with the formatting command, leaving (inadvertantly?)
weak bits in between, and later filling in the sector data with a
regular MC6843 SSW command.  The problem happens when each written
sector is verified (by a SSR command), the controller has to pass over
all these weak bits in a full disk rotation, before coming back to the
targeted sector, with a significant risk a finding a wrong sync code,
and aborting the SSR command. The BIOS code contains workaround code to
handle this case (6 retries max), but the emulation of weak area in mame
generates too many random sync codes in these weak bits.